app/pyproject.toml
[project]
name = "nexus-ark"
version = "0.2.0"
description = "Nexus Ark: AI Persona Interaction System"
authors = [{ name = "Kenomendako", email = "kenomendako@example.com" }]
readme = "README.md"
requires-python = ">=3.11, <3.14"
dependencies = [
    # --- Core Frameworks ---
    "gradio==6.14.0",
    "gradio_client",
    "langchain",
    "langchain-core",
    "langchain-google-genai>=4.2.0",
    "langgraph",
    "google-genai>=1.59.0",
    "google-api-core",
    "pillow",
    # --- RAG (Knowledge Base) Feature ---
    "faiss-cpu",
    "langchain-faiss",
    # langchain-community is often needed for various document loaders
    "langchain-community",
    # --- UI & Data Handling ---
    "pandas==2.3.2",
    "beautifulsoup4==4.13.5",
    "lxml==6.0.2",
    "pypdf==5.1.0",
    # --- Subprocess & Notifications ---
    "psutil==7.1.0",
    "plyer==2.1.0",
    "schedule==1.2.2",
    "requests==2.32.5",
    # --- File & System Utilities ---
    "filetype==1.2.0",
    "filelock>=3.20.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "ijson==3.4.0",
    "ddgs==9.9.1",
    "concurrent-log-handler==0.9.28",
    "send2trash>=1.8.0",
    # --- LangChain/LangGraph Dependencies (Verified) ---
    "aiohttp==3.12.15",
    "anyio==4.11.0",
    "fastapi==0.117.1",
    "httpx==0.28.1",
    "numpy==2.3.3",
    "matplotlib",
    "orjson==3.11.3",
    "pydantic==2.11.9",
    "PyYAML==6.0.2",
    "SQLAlchemy==2.0.43",
    "tenacity==9.1.2",
    # --- Audio dependency ---
    "pydub==0.25.1",
    # --- Markdown rendering ---
    "markdown-it-py~=3.0.0",
    "mdit-py-plugins~=0.4.0",
    # --- Multi-Provider Support ---
    "langchain-openai",
    "langchain-tavily",
    # --- Topic Clustering Feature ---
    "hdbscan>=0.8.33",
    "sentence-transformers>=2.2.0",
    "python-chess",
    "playwright>=1.58.0",
    "pytest-playwright>=0.7.2",
    "tufup>=0.10.0",
    "langchain-anthropic>=1.3.3",
    # --- Screenshot Feature ---
    "mss>=9.0.0",
    "tweepy>=4.16.0",
    "discord.py>=2.5.0",
    "line-bot-sdk",
    "mcp>=1.2.0",
]

[project.optional-dependencies]
win = [
    "pywin32==311 ; sys_platform == 'win32'"
]
local = [
    "llama-cpp-python"
]

[tool.uv]
managed = true
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
